Move Landing and QrScan into separate components and add routing.
